+++ backend/lib/backend_web/controllers/instance_controller.ex
@@ -13,7 +13,7 @@ defmodule BackendWeb.InstanceController do
 
   def show(conn, %{"id" => domain}) do
     instance = Api.get_instance!(domain)
-    last_crawl = get_last_successful_crawl(domain)
+    last_crawl = get_last_crawl(domain)
     render(conn, "show.json", instance: instance, crawl: last_crawl)
   end